+++ mobile/android/android-components/components/feature/session/src/main/java/mozilla/components/feature/session/SessionUseCases.kt
@@ -4,6 +4,7 @@
 
 package mozilla.components.feature.session
 
+import mozilla.components.browser.state.action.ContentAction
 import mozilla.components.browser.state.action.CrashAction
 import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.action.LastAccessAction
@@ -98,6 +99,13 @@ class SessionUseCases(
                     parent = parentEngineSession,
                     flags = flags,
                     additionalHeaders = additionalHeaders,
+                )
+                // Update the url in content immediately until the engine updates with any new changes to the state.
+                store.dispatch(
+                    ContentAction.UpdateUrlAction(
+                        loadSessionId,
+                        url,
+                    ),
                 )
                 store.dispatch(
                     EngineAction.OptimizedLoadUrlTriggeredAction(